refactor: extract shared IOSAlert component to fix SonarCloud duplication (#186)
Extract the iOS-style centered alert Dialog into a reusable IOSAlert
component in src/components/composites/. This eliminates the code
duplication between SettingsScreen (Reset/Import alerts) and
DeleteWordDialog that was flagged by SonarCloud (4.9% > 3% threshold).
SettingsScreen and DeleteWordDialog now import IOSAlert from composites.
